Benchmarks for dbdsqr
This is my first stab at benchmarks for dbdsqr.  I construct a random
bidiagonal matrix using dgebrd, and then optionally check the
performance when VT is used.  I don't have a sense for what typical
matrix sizes should be, so I picked some of the ones from the tests.
I'm also not sure how useful the additional outputs VT, etc. are, so I
might be heading in the wrong direction here.